+++ CHANGELOG.md
@@ -11,7 +11,7 @@
 
 ##### Bug Fixes
 
-- None.
+- Protocol members whose implementation is provided by an external type, yet aren't referenced via a value type are now identified as unused.
 
 ## 2.2.2 (2020-11-23)